# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Royal Jersey BD API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface League {
  id: number;
  name: string;
  slug: string;
  country?: string;
  logoUrl?: string;
  isInternational: boolean;
  sortOrder?: number;
  productCount?: number;
  createdAt?: string;
}

export type ProductEdition =
  (typeof ProductEdition)[keyof typeof ProductEdition];

export const ProductEdition = {
  player: "player",
  fan: "fan",
  kid: "kid",
  premium: "premium",
} as const;

export interface Product {
  id: number;
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  imageUrl: string;
  images?: string[];
  categoryId: number;
  categoryName: string;
  edition: ProductEdition;
  fabricType?: string;
  sizes?: string[];
  inStock: boolean;
  stockCount?: number;
  isFeatured?: boolean;
  isNew?: boolean;
  discountPercent?: number;
  rating?: number;
  reviewCount?: number;
  leagueId?: number;
  leagueName?: string;
  leagueLogoUrl?: string;
  teamName?: string;
  createdAt?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  imageUrl?: string;
  productCount?: number;
}

export interface CartItem {
  productId: number;
  name: string;
  imageUrl: string;
  price: number;
  quantity: number;
  size?: string;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  productId: number;
  quantity: number;
  size?: string;
}

export interface Offer {
  id: number;
  title: string;
  description: string;
  discountPercent?: number;
  imageUrl?: string;
  validUntil?: string;
  code?: string;
  isActive?: boolean;
}

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  card: "card",
  bkash: "bkash",
  nagad: "nagad",
  rocket: "rocket",
  cod: "cod",
} as const;

export interface CreateOrderBody {
  items: CartItem[];
  customerName: string;
  phone: string;
  address: string;
  paymentMethod: CreateOrderBodyPaymentMethod;
  notes?: string;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  total: number;
  paymentMethod?: string;
  createdAt?: string;
}

export interface AdminOrder {
  id: number;
  orderNumber: string;
  status: string;
  total: number;
  paymentMethod?: string;
  customerName: string;
  phone: string;
  address: string;
  notes?: string;
  items?: CartItem[];
  createdAt?: string;
}

export type AdminProductBodyEdition =
  (typeof AdminProductBodyEdition)[keyof typeof AdminProductBodyEdition];

export const AdminProductBodyEdition = {
  player: "player",
  fan: "fan",
  kid: "kid",
  premium: "premium",
} as const;

export interface AdminProductBody {
  name: string;
  description?: string;
  price: number;
  originalPrice?: number;
  imageUrl: string;
  images?: string[];
  categoryId: number;
  edition: AdminProductBodyEdition;
  fabricType?: string;
  sizes?: string[];
  inStock: boolean;
  stockCount?: number;
  isFeatured?: boolean;
  isNew?: boolean;
  discountPercent?: number;
  leagueId?: number;
  teamName?: string;
}

export interface AdminOfferBody {
  title: string;
  description: string;
  discountPercent?: number;
  imageUrl?: string;
  validUntil?: string;
  code?: string;
}

export interface AdminCategoryBody {
  name: string;
  slug: string;
  imageUrl?: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface AdminMe {
  isAdmin: boolean;
  userId?: string;
}

export type ListProductsParams = {
  categoryId?: number;
  edition?: ListProductsEdition;
  search?: string;
  inStock?: boolean;
  leagueId?: number;
  sort?: ListProductsSort;
  minPrice?: number;
  maxPrice?: number;
};

export type ListProductsEdition =
  (typeof ListProductsEdition)[keyof typeof ListProductsEdition];

export const ListProductsEdition = {
  player: "player",
  fan: "fan",
  kid: "kid",
  premium: "premium",
} as const;

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  newest: "newest",
  popular: "popular",
} as const;

export interface SiteSettings {
  [key: string]: string;
}

export interface HeroSlide {
  title: string;
  subtitle: string;
  image: string;
  cta: string;
  link: string;
}

export interface FaqItem {
  question: string;
  answer: string;
}