# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Royal Jersey BD API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /wishlist/{productId}:
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        edition:
          type: string
          enum: [player, fan, kid, premium]
        fabricType:
          type: string
        sizes:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        discountPercent:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        leagueId:
          type: integer
        leagueName:
          type: string
        leagueLogoUrl:
          type: string
        teamName:
          type: string
        createdAt:
          type: string
      required: [id, name, price, imageUrl, categoryId, categoryName, edition, inStock]
...
    CartItem:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        quantity:
          type: integer
        size:
          type: string
      required: [productId, name, imageUrl, price, quantity]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        total:
          type: number
        itemCount:
          type: integer
      required: [items, subtotal, total, itemCount]
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        size:
          type: string
      required: [productId, quantity]
...