# .migration-backup/lib/api-zod/src/generated/types/addToCartBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Royal Jersey BD API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddToCartBody {
  productId: number;
  quantity: number;
  size?: string;
}